Updates: Contact and ContactDTO
- includes attribute id on Contact
- uses inner class Data to hold "dados" field in ContactDTO

Delete: removes ContactInfo class

Add: class Request, RequestDTO and RequestType
