+++ app/src/main/java/br/com/ufv/inf311/solicitaedu/model/RequestType.java
@@ -1,0 +1,4 @@
+package br.com.ufv.inf311.solicitaedu.model;
+
+public enum RequestType {
+}
